add generic write function to be used by write16 and write32
This commit adds a write16 function needed to support some old SoCs.
To reduce the amount of duplicate code, a generic write function is
introduced. It takes a length of a data unit as an argument defaulting
to 32. In the future, this codebase can be extended to support writing
64-bit values.

Making similar changes to read function could also be considered.